requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
djangorestframework==3.14.0
frozendict==2.4.0
html5lib==1.1
idna==3.6
lxml==5.1.0
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.0
peewee==3.17.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.0
webencodings==0.5.1
yfinance==0.2.36
